pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "durc-triage"
version = "0.1.0"
description = "AI-enabled pipeline for triaging dual-use research of concern in preprints"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "structlog>=24.0.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "rapidfuzz>=3.9.0",
    "lxml>=5.2.0",
    "apscheduler>=3.10.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
    "respx>=0.22.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]
